Conflicts:
	5TcpExample/echoserver.c
	head.h

+++ head.h
@@ -24,11 +24,12 @@
 
 #define BUFFSIZE_MY 1024
 
+
 struct args{
-    long arg1;
-    long arg2;
+    int arg1;
+    int arg2;
 };
 
 struct result{
-    long sum;
+    int sum;
 };
